Finished VS Versions page for now on

+++ src/renderer/src/pages/VS/Versions/Versions.svelte
@@ -22,31 +22,55 @@
   import Sheet from '@renderer/lib/ui/components/Sheet.svelte'
   import ScrollableContainer from '@renderer/lib/ui/layout/ScrollableContainer.svelte'
   import Form from '@renderer/lib/ui/components/Form.svelte'
+  import { ask } from '@renderer/lib/ui/app/AlertManager.svelte'
+  import FormInfo from '@renderer/lib/ui/components/FormInfo.svelte'
 
   Breadcrumbs.instance.segments = [{ label: m.vintagestory__versions(), href: '/vs/versions' }]
 
-  let addVersionDialogOpen = $state(false)
+  // If the dialog to install a version is open
+  let isAddVSVersionDialogOpen = $state(false)
 
+  // VS Version form data
   let version: RAPIVSVersion | undefined = $state()
+  let versionErrors: string[] = $state([])
+
+  // Path form data
   let path: string | undefined = $state()
+  let pathErrors: string[] = $state([])
 
   function manageInstallVersion(): void {
-    let missingFields: number = 0
+    let errors = 0
 
-    if (!version) missingFields++
+    // Check if a version is selected
+    if (!version) {
+      versionErrors = ['You need to select a version!']
+      errors++
+    }
 
-    if (missingFields <= 0) {
+    // Install the version if there are no errors
+    if (errors <= 0) {
       version?.addAndInstall(path)
-      addVersionDialogOpen = false
+      clearInstallVersion()
       return
     }
 
+    // If there are errors, show a toast
     const toast = new Toast({
       title: m.toasts__title__missing_fields(),
       type: Toast.Type.DANGER,
       description: m.toasts__description__missing_fields()
     })
     Toasts.instance.addToast(toast)
+  }
+
+  function clearInstallVersion(): void {
+    isAddVSVersionDialogOpen = false
+
+    version = undefined
+    versionErrors = []
+
+    path = undefined
+    pathErrors = []
   }
 </script>
 
@@ -58,7 +82,7 @@
 
         <Tooltip disableHoverableContent>
           {#snippet trigger()}
-            <Button mode="neutral" onclick={() => (addVersionDialogOpen = true)}>
+            <Button mode="neutral" onclick={() => (isAddVSVersionDialogOpen = true)}>
               <PHPlusBoldIcon class="text-2xl text-current/50" />
             </Button>
           {/snippet}
@@ -93,8 +117,11 @@
                         mode="danger"
                         disabled={vsVersion.state === VSVersion.State.DELETING || vsVersion.state === VSVersion.State.INSTALLING}
                         onclick={async () => {
-                          await vsVersion.delete()
-                          Data.instance.vsVersions.delete(vsVersion)
+                          const ok = await ask('Are you sure?', 'Are you sure you want to delete this VS Version? This action cannot be undone!')
+                          if (ok) {
+                            await vsVersion.delete()
+                            Data.instance.vsVersions.delete(vsVersion)
+                          }
                         }}
                       >
                         <PHTrashBoldIcon />
@@ -133,7 +160,14 @@
   </FlexContainer>
 </ScrollableContainer>
 
-<Sheet bind:open={addVersionDialogOpen} width="xs" title="Install VS Version" description="Select the VS Version you want to install and click install.">
+<Sheet
+  bind:open={isAddVSVersionDialogOpen}
+  contentProps={{ onEscapeKeydown: clearInstallVersion, onInteractOutside: clearInstallVersion }}
+  onclose={clearInstallVersion}
+  width="xs"
+  title="Install VS Version"
+  description="Select the VS Version you want to install and click install."
+>
   <Form onsubmit={manageInstallVersion}>
     <FlexContainer direction="col" height="full" gap="xl" alignY="between">
       <FlexContainer direction="col" gap="base">
@@ -143,7 +177,11 @@
             <Info>{m.descriptions__vs_version_to_install()}</Info>
           </FlexContainer>
 
-          <VersionsToInstall bind:version required />
+          <VersionsToInstall bind:version mode={versionErrors.length > 0 ? 'danger' : 'neutral'} onValueChange={() => (versionErrors = [])} required />
+
+          <FlexContainer direction="col" gap="xs">
+            <FormInfo error={versionErrors} />
+          </FlexContainer>
         </FlexContainer>
 
         <FlexContainer direction="col" gap="sm">
@@ -167,11 +205,15 @@
 
             <Input type="text" name={m.labels__vs_version_path()} placeholder={m.labels__vs_version_path()} value={path} readonly required />
           </FlexContainer>
+
+          <FlexContainer direction="col" gap="xs">
+            <FormInfo error={pathErrors} />
+          </FlexContainer>
         </FlexContainer>
       </FlexContainer>
 
       <FlexContainer gap="sm">
-        <Button mode="neutral" width="flex-1" onclick={() => (addVersionDialogOpen = false)}>
+        <Button mode="neutral" width="flex-1" onclick={clearInstallVersion}>
           {m.common__cancel()}
         </Button>
